telling bots to not index default config page
Because default config is a redirect back to the main page, Google throws indexing errors on it. Now that it is listed in robots.txt as a page to not index, Google is throwing errors about attempting to index it even though it's excluded in robots.txt. So, adding a no-folow tag to the link to remove indexing the defaultconfig/ page.